Built program capable of generating and executing ML model for predicting pIC50 values against a specified target protein / disease
Gathers and parses data from the ChEMBL website and processes data on a chosen target protein or disease, ultimately obtaining a fingerprint of molecular descriptors for each compound as well as its pIC50 (potency metric) value. It then builds a machine learning model based on that data and allows for predictions for any compound against the initially investigated target disease or protein.